refactor(web): extract shared packet table and inspector into core
Promote the packet table and packet detail inspector out of the pdump
module into the shared web core so other packet-display surfaces can
reuse them as generic, capture-agnostic building blocks. The pdump page
now adapts its captured packets onto the shared components and supplies
its capture metadata through the inspector's optional section, keeping
the existing collapsible layout.
