+++ web/src/core/components/packet/index.ts
@@ -1,0 +1,6 @@
+export { PacketTable } from './PacketTable';
+export type { PacketTableProps } from './PacketTable';
+export { default as PacketInspector, Section, KV } from './PacketInspector';
+export type { PacketInspectorProps } from './PacketInspector';
+export type { PacketRow, PacketSortColumn, PacketSortDirection, PacketSortState } from './types';
+export { PKT_ROW_HEIGHT, PKT_HEADER_HEIGHT, PKT_FOOTER_HEIGHT, PKT_SEARCH_BAR_HEIGHT, PKT_OVERSCAN, PKT_TOTAL_WIDTH, pktCellStyles } from './constants';
